added helpful var docs to the routes files changed container get console kernel to console application refactored the tests
