+++ src/Resource/web.php.stub
@@ -5,6 +5,8 @@ use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Viserio\Component\Http\Response\HtmlResponse;
 
+/** @var \Viserio\Component\Routing\Router $router */
+
 $router->get('/', function (ServerRequestInterface $request, array $data, ContainerInterface $container) {
     $view = $container->get('view');
